+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - AVRCP: shorten default SDP Service and Provider Names
 - GATT Client: emit query complete event for gatt_client_discover_characteristic_descriptors in next run loop iteration
 - HFP: report 1.9 in SDP record
+- HID Parser: ignore Report Items without Usage
 - btstack_crypto: allow MBEDTLS config via MBEDTLS_CONFIG_FILE
 - remove old Zephyr 1.9 port